[TAN-3636] Renamed SuccessModal (spelt incorrectly)
